pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "cjber", email = "cjberragan@gmail.com"},
]
dependencies = [
    "llama-index>=0.9.15",
    "llama-hub>=0.0.58",
    "pinecone-client>=2.2.4",
    "unstructured[all-docs]>=0.11.2",
    "requests>=2.31.0",
    "fastapi>=0.105.0",
    "uvicorn>=0.24.0.post1",
    "streamlit>=1.29.0",
    "polars>=0.19.19",
    "more-itertools>=10.1.0",
    "ipython>=8.18.1",
    "pre-commit>=3.6.0",
    "ipdb>=0.13.13",
    "llama-cpp-python>=0.2.23",
]
requires-python = ">=3.11,<3.12"
readme = "README.md"
license = {text = "MIT"}

